# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Employee Attrition Risk Prediction API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Employee {
  id: number;
  employeeId: string;
  name: string;
  age: number;
  department: string;
  jobRole: string;
  yearsAtCompany: number;
  monthlyIncome: number;
  attritionRisk: number;
  riskLevel: string;
  attrited: boolean;
  jobSatisfaction: number;
  workLifeBalance: number;
  overTime: boolean;
}

export interface EmployeeDetail {
  id: number;
  employeeId: string;
  name: string;
  age: number;
  department: string;
  jobRole: string;
  yearsAtCompany: number;
  monthlyIncome: number;
  attritionRisk: number;
  riskLevel: string;
  attrited: boolean;
  jobSatisfaction: number;
  workLifeBalance: number;
  overTime: boolean;
  education: number;
  distanceFromHome: number;
  numCompaniesWorked: number;
  totalWorkingYears: number;
  trainingTimesLastYear: number;
  yearsInCurrentRole: number;
  yearsSinceLastPromotion: number;
  environmentSatisfaction: number;
  jobInvolvement: number;
  performanceRating: number;
  stockOptionLevel: number;
  maritalStatus: string;
  gender: string;
}

export interface EmployeeListResponse {
  employees: Employee[];
  total: number;
  page: number;
  pageSize: number;
  totalPages: number;
}

export interface AnalyticsSummary {
  totalEmployees: number;
  attritionCount: number;
  attritionRate: number;
  highRiskCount: number;
  mediumRiskCount: number;
  lowRiskCount: number;
  avgAttritionRisk: number;
  avgMonthlyIncome: number;
  avgYearsAtCompany: number;
}

export interface FeatureImportanceItem {
  feature: string;
  importance: number;
  direction: string;
}

export interface DepartmentBreakdown {
  department: string;
  total: number;
  attrited: number;
  highRisk: number;
  attritionRate: number;
  avgRisk: number;
}

export interface RiskDistributionItem {
  riskLevel: string;
  count: number;
  percentage: number;
}

export interface ModelMetrics {
  accuracy: number;
  f1Score: number;
  rocAuc: number;
  precision: number;
  recall: number;
  cvMean: number;
  cvStd: number;
  trainSize: number;
  testSize: number;
}

export interface PredictionInput {
  age: number;
  monthlyIncome: number;
  yearsAtCompany: number;
  jobSatisfaction: number;
  workLifeBalance: number;
  overTime: boolean;
  distanceFromHome: number;
  numCompaniesWorked: number;
  totalWorkingYears: number;
  yearsInCurrentRole: number;
  yearsSinceLastPromotion: number;
  environmentSatisfaction: number;
  jobInvolvement: number;
  performanceRating: number;
  stockOptionLevel: number;
  trainingTimesLastYear: number;
}

export interface PredictionResult {
  attritionRisk: number;
  riskLevel: string;
  confidence: number;
  topFactors: FeatureImportanceItem[];
}

export type ListEmployeesParams = {
  page?: number;
  pageSize?: number;
  /**
   * @nullable
   */
  department?: string | null;
  /**
   * @nullable
   */
  riskLevel?: string | null;
  /**
   * @nullable
   */
  search?: string | null;
};