# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Employee Attrition Risk Prediction API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /employees:
    get:
      operationId: listEmployees
      tags: [employees]
      summary: List employees
      description: Get paginated list of employees with attrition predictions
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            default: 20
        - name: department
          in: query
          schema:
            type: ["string", "null"]
        - name: riskLevel
          in: query
          schema:
            type: ["string", "null"]
        - name: search
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: Employee list with pagination
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmployeeListResponse"
  /employees/{id}:
    get:
      operationId: getEmployee
      tags: [employees]
      summary: Get employee details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Employee details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmployeeDetail"
        "404":
          description: Not found
...
components:
  schemas:
...
    Employee:
      type: object
      properties:
        id:
          type: integer
        employeeId:
          type: string
        name:
          type: string
        age:
          type: integer
        department:
          type: string
        jobRole:
          type: string
        yearsAtCompany:
          type: number
        monthlyIncome:
          type: number
        attritionRisk:
          type: number
        riskLevel:
          type: string
        attrited:
          type: boolean
        jobSatisfaction:
          type: integer
        workLifeBalance:
          type: integer
        overTime:
          type: boolean
      required:
        - id
        - employeeId
        - name
        - age
        - department
        - jobRole
        - yearsAtCompany
        - monthlyIncome
        - attritionRisk
        - riskLevel
        - attrited
        - jobSatisfaction
        - workLifeBalance
        - overTime
    EmployeeDetail:
      type: object
      properties:
        id:
          type: integer
        employeeId:
          type: string
        name:
          type: string
        age:
          type: integer
        department:
          type: string
        jobRole:
          type: string
        yearsAtCompany:
          type: number
        monthlyIncome:
          type: number
        attritionRisk:
          type: number
        riskLevel:
          type: string
        attrited:
          type: boolean
        jobSatisfaction:
          type: integer
        workLifeBalance:
          type: integer
        overTime:
          type: boolean
        education:
          type: integer
        distanceFromHome:
          type: integer
        numCompaniesWorked:
          type: integer
        totalWorkingYears:
          type: number
        trainingTimesLastYear:
          type: integer
        yearsInCurrentRole:
          type: number
        yearsSinceLastPromotion:
          type: number
        environmentSatisfaction:
          type: integer
        jobInvolvement:
          type: integer
        performanceRating:
          type: integer
        stockOptionLevel:
          type: integer
        maritalStatus:
          type: string
        gender:
          type: string
      required:
        - id
        - employeeId
        - name
        - age
        - department
        - jobRole
        - yearsAtCompany
        - monthlyIncome
        - attritionRisk
        - riskLevel
        - attrited
        - jobSatisfaction
        - workLifeBalance
        - overTime
        - education
        - distanceFromHome
        - numCompaniesWorked
        - totalWorkingYears
        - trainingTimesLastYear
        - yearsInCurrentRole
        - yearsSinceLastPromotion
        - environmentSatisfaction
        - jobInvolvement
        - performanceRating
        - stockOptionLevel
        - maritalStatus
        - gender
    EmployeeListResponse:
      type: object
      properties:
        employees:
          type: array
          items:
            $ref: "#/components/schemas/Employee"
        total:
          type: integer
        page:
          type: integer
        pageSize:
          type: integer
        totalPages:
          type: integer
      required:
        - employees
        - total
        - page
        - pageSize
        - totalPages
...