# lib/api-zod/src/generated/types/employee.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Employee Attrition Risk Prediction API
 * OpenAPI spec version: 0.1.0
 */

export interface Employee {
  id: number;
  employeeId: string;
  name: string;
  age: number;
  department: string;
  jobRole: string;
  yearsAtCompany: number;
  monthlyIncome: number;
  attritionRisk: number;
  riskLevel: string;
  attrited: boolean;
  jobSatisfaction: number;
  workLifeBalance: number;
  overTime: boolean;
}